REFACTOR[#55]: refatoracao da tela details_screen
