+++ .packages
@@ -3,32 +3,32 @@
 # 
 # For more info see: https://dart.dev/go/dot-packages-deprecation
 # 
-# Generated by pub on 2022-06-15 16:12:37.460540.
-async:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.8.2/lib/
-boolean_selector:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/lib/
-characters:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.2.0/lib/
-charcode:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/charcode-1.3.1/lib/
-clock:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/clock-1.1.0/lib/
-collection:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.16.0/lib/
-cupertino_icons:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-1.0.5/lib/
-fake_async:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/fake_async-1.3.0/lib/
-flutter:file:///home/hiltonthallyson/snap/flutter/common/flutter/packages/flutter/lib/
-flutter_lints:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_lints-1.0.4/lib/
-flutter_test:file:///home/hiltonthallyson/snap/flutter/common/flutter/packages/flutter_test/lib/
-lints:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/lints-1.0.1/lib/
-matcher:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.11/lib/
-material_color_utilities:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.1.4/lib/
-meta:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.7.0/lib/
-nested:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/nested-1.0.0/lib/
-path:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/path-1.8.1/lib/
-percent_indicator:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/percent_indicator-4.2.2/lib/
-provider:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/provider-6.0.3/lib/
-sky_engine:file:///home/hiltonthallyson/snap/flutter/common/flutter/bin/cache/pkg/sky_engine/lib/
-source_span:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.8.2/lib/
-stack_trace:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/lib/
-stream_channel:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/lib/
-string_scanner:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.0/lib/
-term_glyph:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.0/lib/
-test_api:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.9/lib/
-vector_math:file:///home/hiltonthallyson/snap/flutter/common/flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.2/lib/
+# Generated by pub on 2022-06-15 18:48:42.346305.
+async:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.8.2/lib/
+boolean_selector:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/lib/
+characters:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.2.0/lib/
+charcode:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/charcode-1.3.1/lib/
+clock:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/clock-1.1.0/lib/
+collection:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.16.0/lib/
+cupertino_icons:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-1.0.5/lib/
+fake_async:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/fake_async-1.3.0/lib/
+flutter:file:///D:/Programas/Development/sdks/flutter/packages/flutter/lib/
+flutter_lints:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_lints-1.0.4/lib/
+flutter_test:file:///D:/Programas/Development/sdks/flutter/packages/flutter_test/lib/
+lints:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/lints-1.0.1/lib/
+matcher:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.11/lib/
+material_color_utilities:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/material_color_utilities-0.1.4/lib/
+meta:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.7.0/lib/
+nested:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/nested-1.0.0/lib/
+path:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/path-1.8.1/lib/
+percent_indicator:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/percent_indicator-4.2.2/lib/
+provider:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/provider-6.0.3/lib/
+sky_engine:file:///D:/Programas/Development/sdks/flutter/bin/cache/pkg/sky_engine/lib/
+source_span:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.8.2/lib/
+stack_trace:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/lib/
+stream_channel:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/lib/
+string_scanner:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.0/lib/
+term_glyph:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.0/lib/
+test_api:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.9/lib/
+vector_math:file:///D:/Programas/Development/sdks/flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.2/lib/
 catalogo_filmes:lib/
